mgr/dashboard: RbdMirroringService test suite fails in dev mode
If you have set DEV in unit-test-configuration.ts to true, in order to
run all unit tests much faster, the unit test suite of rbd mirroring
service fails.

The reason is, that it's expecting a call, that is triggered inside the
constructor, which is only done once in for the copy used in the fast
testing mode (if it would be done on within ngOnInit this wouldn't be a
problem).

Fixes: http://tracker.ceph.com/issues/37841
Signed-off-by: Stephan Müller <smueller@suse.com>