common/ceph_argparse: make parse_ip_port_vec handle list of addrs or addrvecs
This helper is only used for mon_host.

We want to be able to list addrs or addrvecs.  It is slight wonky because
you could parse something like

 "1.2.3.4,5.6.7.8"

as either a list of addrs or a list of addrvecs.  Since the addr parse
takes a default type, it is preferred, so first try parsing as an addr
before proceeding.  Addrvecs that are size >1 must have brackets, so we
will always parse "[1.2.3.4,5.6.7.8]" unambiguously (since "[1.2.3.4"
won't parse as an addr).

Signed-off-by: Sage Weil <sage@redhat.com>