mon/MonClient: be less aggressive/paranoid about reopening mon connections
Previously we were reopening any time the name of our mon changed.  That
was pretty much every time, since we started out with names from the
seed monmap like 'noname-[abc]' that never match the real mons.

Instead, only reopen if the addr we are connecting to is no longer in the
monmap.  We don't actually care about the mon's name.

Fixes: http://tracker.ceph.com/issues/37868
Signed-off-by: Sage Weil <sage@redhat.com>