crimson/net: fix unittest_seastar_messenger failure
unittest_seastar_messenger is failed since updated the seastar repo.
std::system_error (error system:22, Invalid argument) will be raised
from `listener->accept()`, which will be re-raised and not be captured.

This fix is to print error logs for the unexpected exceptions in
`SocketMessenger::start()` and not raise them again.

Signed-off-by: Yingxin Cheng <yingxincheng@gmail.com>