common/ip_addr: do not support numa preference if WITH_SEASTAR
in pick_address.cc, set_val_or_die() is called to set the "public_addr"
and/or "cluster_addr". but since the sharded ConfigProxy cannot
set_val_or_die() synchronously, we cannot do it this way.

we need to fix pick_addresses() so it works with seastar later. either
by making the involved function return future<>, or making the set_val()
call non-blocking or synchronous if we can go with 1:1 mapping
threading model.

but at this moment, let's just disable this feature in crimson-osd.

Signed-off-by: Kefu Chai <kchai@redhat.com>