mon/MDSMonitor: skip offline ERR for down fs
The idea here is that a file system marked down is considered intentionally
archived by the sysadmin. Consider the future use-case where there are dozens
or hundreds of file systems in use by dynamic applications that come and go.  A
file system may only be brought up when the corresponding application is
online.

Fixes: http://tracker.ceph.com/issues/20611

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>