scripts: add cli tools for reimaging nodes without locking
Add teuthology-reimage cli tool to be able to provision nodes using
Fog or Pelagos without locking and unlocking.

This is useful, for example, when someone just locks the node for
development or debugging purposes and do not want to release while
resetting the image, because it can happen that there are no free
nodes available.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>
Signed-off-by: Vasu Kulkarni <vasu@redhat.com>