rgw: tests: initial scripts to test scheduler
This is a basic python3 only script that can fire a specified count of requests
and an optional payload(currently only a fixed buffer) against a url. This is
built atop of python3 asyncio and will only work on pythons > 3.5 (which should
be default py3 in most distros). We create a single session per url and fire off
the requested number of requests as coroutines which py3 usually schedules as
coroutines. Right now the most basic form is just counting the response status
though any sort of response handler can be written for more complex use cases.
Currently exponential backoff is not yet implemented for 503s. A
basic configuration is attached.

With the current configuration a test run would create an output like:
```sh
$ python test.py test.conf
            {'GET': {'/': Counter({503: 499, 200: 1}),
                     '/versiontest': Counter({503: 470, 200: 30})},
             'PUT': {'/versiontest': Counter({503: 484, 200: 16}),
                     '/versiontest/obj1': Counter({503: 499, 200: 1})}})
```

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>