osd: refuse to start if release > recorded min_osd_release + 2
If we try to start up the objectstore, we may make writeable changes to
(say) rocksdb that are not backwards compatible.  This happens, for
example, if you start a mimic osd.  Even if the compatset checks fail,
rocksdb may have written something that is not backwards compatible.

Fixes: http://tracker.ceph.com/issues/38076
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 9f7713a905d67441b28371e4494e9447319d2129)

Conflicts:
	src/ceph_osd.cc
- include common/version.h for ceph_release()
- use exit instead of forker.exit