msg/async/ProtocolV2: respect policy.resetcheck on cookie==0
It's possible the accepting end knows the cookie but the connecting end
does not if the fault happens before it receives server_ident.  If the
next connection attempt is in the other direction, we need to avoid a
reset or the queued messages at the first site will get lost.

Fixes: http://tracker.ceph.com/issues/38118
Signed-off-by: Sage Weil <sage@redhat.com>