mgr: remove _ceph_get_module_option_ex() method
It seems that the current implementation has some side effects which results in segfaults in the Mgr. It looks like boxing parameters for feeding them into a Python function written in C++ does not work very well or we did not handle the GIL correctly here. Because of that the old code from the ceph_(get|set)_module_option functions are used again (see #25651). Thus we have some lines of more or less duplicate code, but we get rid of the problems that arise from the current implementation.

Previous improvements like #26112, #26240 and #26149 are still useful for the ceph_(get|set)_module_option_ex functions.

Signed-off-by: Volker Theile <vtheile@suse.com>