mgr/dashboard: use orchestrator_cli to figure out the backend
Currently the dashboard has its own setting for the orchestrator
backend, which means that one has to set it in both the dashboard and
the orchestrator_cli module if you intend to use both.

Change the dashboard code to use the setting in the orchestrator_cli
module, and eliminate the dashboard-specific setting.

Signed-off-by: Jeff Layton <jlayton@redhat.com>