mon/Monitor: do not join cluster that is >2 releases old
This enforces the N+2 upgrade rule from the mon's perspective.

Note that this safety check is not as safe as the OSDs.  Notably, we
start up our backend store (rocksdb) *before* we probe other monitors
and discover any newer monmap that tells us we shouldn't join.  If there
is a *rocksdb* backward-compatibility problem it is too late by this
point.  Unfortunately, I don't see an easy way to get this far before
rocksdb is read-write--not without a lot more code, at least!

However, we'll still protect against a whole class of other potential
problems by not getting involved in a cluster that is too old.  :)

Signed-off-by: Sage Weil <sage@redhat.com>