restful: Fix regression when traversing leaf nodes
The commit 23b6c90 introduced a regression when traversing leaf nodes.
The issue is that it traverses the keys of a `dict` returned by
`nodes_by_id`, not the actual `items` of the node. That resulted in an
500 error because it tried to treat `str` as a `dict` and failed.

Signed-off-by: Boris Ranto <branto@redhat.com>
(cherry picked from commit fb51bad963cc649fb12ca0ed8b738a2e819299b6)