msg/async/ProtocolV2: assume peer has SERVER_NAUTILUS for encoding v2 frame
The v2 frames may contain entity_addr_t and entity_addrvec_t.  Those have
been understood since the MSG_ADDR2 feature, but with nautilus we've
additionally added the TYPE_ANY, and we encode TYPE_ANY addrs as
TYPE_LEGACY when the peer lacks SERVER_NAUTILUS.

For the purposes of the v2 protocol, we assume the peer understands
TYPE_ANY.

No other complex types are used in the v2 frame protocol, so this has no
effect on, e.g., what features we assume the peer has for the purposes
of encoding actual messages.

Signed-off-by: Sage Weil <sage@redhat.com>