librbd: skip mirror image verification when pruning rbd-mirror watchers
When an image is moved to the trash, mirroring is disabled which will remove
the image id from the mirroring directory. If you then quickly attempt to
remove the image from the trash, it might fail if the rbd-mirror daemon hasn't
closed the image since it fails to filter out rbd-mirror daemon watchers
against images with mirroring disabled.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>