osd: do not send peers really old maps
We may receive a message that sat in a queue for a while with a low
priority and is tagged with an older epoch.  Don't send a bunch of old
maps that we have already sent the peer.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit fa58979c2ddf2f4aa68386d04a2dddae0cf38c7a)