rgw: Refactor Aio and AioThrottle
To support in-development work on local caching (and in-development
work on having RGW use the 'unleashed' version of RADOS) divorce the
implementation of AioThrottle from the specifics of librados in accord
with a design that Casey Bodley <cbodley@redhat.com> outlined.

Also include a 'support' function to simplify using the aio_operate
machinery on librados. It's a bit ugly given how completions are
implemented, but all that's hidden away.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>