CLI: ability to change file ownership
When creating/fetching key it's nice to have the ability to change the
ownership of the created file.

This commit adds the '--setuser' and 'setgroup' which respectively apply
the desired owner and group to a file user when '--output' is passed.

Closes: https://tracker.ceph.com/issues/38370
Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 0e26090960213ef911cfc7bfcdf2925ddf1b268f)

Conflicts:
	src/ceph.in : Resolved in  new_style_command