rbd-mirror: complete pool watcher initialization if object missing
If the mirroring object is missing, complete the initialization and
continue to retry in the background. This is useful for cases where
the remote doesn't (yet) have mirroring enabled but the remote
pool watcher initialization is delaying the leader watcher promotion
to the point where the leader is blacklisted by its peers.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 80954cd914c86e11ffb6a8cbfcb21202cb8131b5)