rgw admin: implement a lc fix option
An radosgw-admin lc fix --bucket <> option is added which checks if the bucket
entry exists in the corresponding lc shard and creates it if not. In case of
resharded buckets not running a fixed rgw that writes/compares the marker this
would write a new entry with the marker as the old entry would've already been
deleted by a LC process. We currently don't cleanup the stale entry as it is
assumed this would be picked up by the LC processor already or would be picked
up in the next cycle.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>
(cherry picked from commit 4528db4b642c37f67b3338ce04979b1ba7772e3e)