mon/OSDMonitor: use legacy_equals to compare osd addrs
The OSDs may bind to v2+v1 due to ms_bind_msgr2, but we may only use the
v1 address in the OSDMap due to

 - the osd->mon connection being v1 (mon isn't using v2 yet)
 - the osd require_osd_release < nautilus
 - ???

Conversely, the OSD may end up with a v2+v1 address in the osdmap, but
may connect to the current monitor via v1 only, due to the mon only being
bound to a v1 address.

If/when this happens, make sure we are doing a loose comparison of the
osdmap addrs vs the osd's addr they are connecting from.

Fixes: http://tracker.ceph.com/issues/38598
Signed-off-by: Sage Weil <sage@redhat.com>