osd/osd_types: remove copy ctor of osd_reqid_t
do not define the copy constructor of osd_reqid_t explicitly. compiler
will define a default one for us. the implicitly defined one will call
the copy constructor of each member variable. so it is as good as the
user-defined one.

another reason to remove this copy ctor is that, it prevents the
compiler from defining a default move constructor, without which, we
cannot return `future<object_info_t>`. as `object_info_t` contains an
instance of `osd_reqid_t`.

Signed-off-by: Kefu Chai <kchai@redhat.com>