crimson/osd: do not pass ceph options to seastar
the program_option parser used by seastar::app_template does not allow
unrecognized options. but ceph options can be specfied by being passed
to ceph applications as command line options, for instance, we can
specify the "key" or "keyfile" when creating an objectstore using
"--mkfs", like:

  ceph-osd --mkfs --key <key>

in this change, all options known by seastar's app_template are
enumerated and stored into a separated vector. so it can be passed to
"app". and the unknown ones are passed to ceph functions.

Signed-off-by: Kefu Chai <kchai@redhat.com>