mon/OSDMonitor: apply osd_crush_update_weight_set for reweight, create-or-move
Since CrushWrapper no longer applies this setting at a low level,
where it can't tell what the real intention is, we instead apply
it at the top command level where we do.

Specifically, we use it to control whether the weight-set weights
are set for the commands

 osd crush reweight
 osd crush reweight-subtree
 osd crush move
 osd crush create-or-move

Note that this (indirectly) affects the way weight-set weights
are initialized for newly created OSDs, since those are added to
the crush map via the 'osd crush create-or-move' command.

Signed-off-by: Sage Weil <sage@redhat.com>