rgw: keystone: options to separately size or disable the keystone secret cache
A new configuration option "rgw_keystone_secret_cache_size" is provided to
control the size of the secret cache independently of the keystone cache.
Because users can easily make multiple keystone tokens (in some cases
a new token is created on every access), the keystone cache may need
a fairly large size in order to be effective.  The secret cache never
needs to be larger than the number of active users.

When "rgw_keystone_secret_cache_size" is set to zero the secret
cache is entirely disabled.  This forces rgw to always use the token
validation logic that used to be the only way to operate.  Sites that
are particularly security conscious may require this mode of operation.
This option may be changed dynamically at runtime without restarting rgw.

Sites that are particularly concerned about security may want to provably
and entirely remove the logic that fetches secret keys from keystone.
The compile-time option "ENABLE_KEYSTONE_SECRET_CACHE=OFF" will remove
the code that uses the secret cache at build time.  Some parts of the
secret cache are left in order to simplify the ifdef logic.

Signed-off-by: Marcus Watts <mwatts@redhat.com>