msg/async/ProtocolV[12]: add ms_learn_addr_from_peer
Optionally learn our address based on what our peer (usually the mon) thinks we are
connecting from, as opposed to what our local socket used for the outbound connection.
This enables us to detect what our NATed address is (as a client) in those environments.

Default to true to match the v1 behavior.  This is a net change in
behavior for v2.

Signed-off-by: Sage Weil <sage@redhat.com>