mon/ConfigMonitor: kill improper assert
otherwise it can be reliably triggerred by the following command:

  ceph config reset 0

There are two ways to fix this problem:
- add some more santicy checks and prevent user from passing in
  any kind of non-positive (not permitted) value
- kill the assert instead

Option 2 sounds to be the preferred choice for me since "0" in general
is the natural representation of "nothing" and hence could always
be used to revert everything to its default or initiative state.

But perhaps I am the only one?

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>