mgr/dashboard: Queue notifications as default
Every time "show" is triggers now the notification gets queued before
sending it out. The notification service will make sure not to send out
duplicated notifications and to combine notifications that have the same
title. The combination of notification is useful for error messages.

For example if the mgr goes down you will get a view toasties notifying
about a 500 error. With the new implementation you get only one with the
different API paths it couldn't get.

Fixes: https://tracker.ceph.com/issues/39034
Signed-off-by: Stephan Müller <smueller@suse.com>