mgr/pg_autoscaler: apply bias to pg_num selection
This is a relatively naive way to apply the bias: we just multiply it
to whatever we would have chosen.  A more clever approach would be to
factor this into the overall cluster-wide PG budget, so that biasing one
pool's PGs up would put downward pressure on other pools.  That is
significantly more complicated, however, and (I think) not worth the
effort.

Signed-off-by: Sage Weil <sage@redhat.com>