vstart: generate environment script suitable for sourcing
At the end, vstart.sh will print out some commands to set up the
environment variables suitably for running things out of the tree.
Unfortunately, they just go to stdout so it's hard to refer to them from
other shells.

Have it tee those lines to a separate vstart_environment.sh file that
can be sourced into the local environment.

While we're at it, let's not use relative paths here, as that makes
the validity of these paths depend on the cwd.

Signed-off-by: Jeff Layton <jlayton@redhat.com>