mgr/DaemonServer: be more conservative when approving safe-to-destroy
by tracing back to any degraded PGs that ever stored on those
specific osds.

However, this is still not very accurate, e.g., we could:
- out an osd say osd.1, wait until all PGs become active+clean
- then out yet another osd (osd.2), check whether osd.1 is
  safe-to-destroy immediately

Step 2 will probably fail if osd.2 happens to share some PGs
(through crush) with osd.1.

In other words, we might become over-conservative when try to approve
a safe-to-destroy request, but I think it should be fine since we still
provide the "--force" option to allow user to arbitrarily destroy an osd?

Fixes: http://tracker.ceph.com/issues/38930
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>