mgr/dashboard: Removes distracting search behavior
The behavior was to not search anything if the search included a
subset of a column name, with the thought in mind that the user would
like to search a specific column for a value.

Now the column name is only ignored after you typed in ':' in order to see
the data in the table again before specifying a more concrete value to
search for.

Fixes: https://tracker.ceph.com/issues/37701
Signed-off-by: Stephan Müller <smueller@suse.com>