mgr/volumes: add mds-monitor module
... that automatically create MDS separate from volumes creation.
This kind of preempts the job of Rook. The idea is that the mgr
can see if the cluster is degraded (e.g. a new fs was just created)
and spin up an MDS on demand. The next step (for future work) would
be allowing such a module to spin up new MDSs in response to heavy load.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>