osd/tier: avoid stale primar pg flush
When flush clone object, we will generate delete op, this uesd to let
head in base pool make clone correctly. but the delete op is send from
tier pg to base pg, so if tier pg partition happen, it will still send
to base pg. so we need extra info to avoid the case.

Detail:

Let's say tier pg has a partited one called pg[1,2]p1, p1 is disconnected
with others and monitor, so it will not receive new osdmap, and keep
active state. the auth pg is already become pg[2,3]p2 ,and start to flush.

When osd.1 reconnect and before advance to new osdmap, the agent_work
may start to flush, and send delete and copy_from to base pool, delete
will be handle and remove head, but copy_from will fail, because the flush
already occured.. and the head in base pool will be deleted wrongly and
can not recovery.

Signed-off-by: Zengran Zhang <zhangzengran@sangfor.com.cn>