mgr/dashboard: Do not default to 'admin' as Admin Resource
The current if-statement would always default to 'admin' as the admin
resource regardless of what 'mgr/dashboard/RGW_API_ADMIN_RESOURCE' is set
to.

In some cases users change this setting in the RGW for various reasons.

The dashboard currently does not support this and will cause the RGW management
not to work.

Fixes: http://tracker.ceph.com/issues/39338

Signed-off-by: Wido den Hollander <wido@42on.com>