rgw: rest_conn: parse_json bails on failed parsing
Currently the json parser parsing errors get ignored while we check the value as
integer reading from a bool. This causes failures in cases when for eg. XML is
supplied to the parser getting ignored and we fail later in the stack with the
default initialized value of T. Also drop the ignored ceph context

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>